+++ tests/xml/mod.rs
@@ -5468,6 +5468,66 @@ mod contain {
     }
 
     #[test]
+    fn contain_paint_block_contains_floats__border_box_ltr() {
+        crate::run_xml_test("contain", "contain_paint_block_contains_floats__border_box_ltr");
+    }
+
+    #[test]
+    fn contain_paint_block_contains_floats__content_box_ltr() {
+        crate::run_xml_test("contain", "contain_paint_block_contains_floats__content_box_ltr");
+    }
+
+    #[test]
+    fn contain_paint_block_contains_floats__border_box_rtl() {
+        crate::run_xml_test("contain", "contain_paint_block_contains_floats__border_box_rtl");
+    }
+
+    #[test]
+    fn contain_paint_block_contains_floats__content_box_rtl() {
+        crate::run_xml_test("contain", "contain_paint_block_contains_floats__content_box_rtl");
+    }
+
+    #[test]
+    fn contain_paint_block_margin_collapse__border_box_ltr() {
+        crate::run_xml_test("contain", "contain_paint_block_margin_collapse__border_box_ltr");
+    }
+
+    #[test]
+    fn contain_paint_block_margin_collapse__content_box_ltr() {
+        crate::run_xml_test("contain", "contain_paint_block_margin_collapse__content_box_ltr");
+    }
+
+    #[test]
+    fn contain_paint_block_margin_collapse__border_box_rtl() {
+        crate::run_xml_test("contain", "contain_paint_block_margin_collapse__border_box_rtl");
+    }
+
+    #[test]
+    fn contain_paint_block_margin_collapse__content_box_rtl() {
+        crate::run_xml_test("contain", "contain_paint_block_margin_collapse__content_box_rtl");
+    }
+
+    #[test]
+    fn contain_paint_flex_baseline__border_box_ltr() {
+        crate::run_xml_test("contain", "contain_paint_flex_baseline__border_box_ltr");
+    }
+
+    #[test]
+    fn contain_paint_flex_baseline__content_box_ltr() {
+        crate::run_xml_test("contain", "contain_paint_flex_baseline__content_box_ltr");
+    }
+
+    #[test]
+    fn contain_paint_flex_baseline__border_box_rtl() {
+        crate::run_xml_test("contain", "contain_paint_flex_baseline__border_box_rtl");
+    }
+
+    #[test]
+    fn contain_paint_flex_baseline__content_box_rtl() {
+        crate::run_xml_test("contain", "contain_paint_flex_baseline__content_box_rtl");
+    }
+
+    #[test]
     fn contain_size_block__border_box_ltr() {
         crate::run_xml_test("contain", "contain_size_block__border_box_ltr");
     }